+++ WebApiCaracterizacion/ModelsMineria/PromediosVolumenProduccionOR.cs
@@ -4,6 +4,6 @@
     {
         public string tipo_plantilla { get; set; }
         public string municipio { get; set; }
-        public decimal promedio { get; set; }
+        public double promedio { get; set; }
     }
 }
